Fix printing of signed numbers
If any operand of the '?' operator is of an unsigned type, the result
is unsigned by default. Thanks to Julian Stecklina for finding
this out.

Fixes #189.

+++ base/src/base/console/console.cc
@@ -276,7 +276,7 @@ void Console::vprintf(const char *format, va_list list)
 				case Format_command::LONG:
 
 					numeric_arg = (cmd.type == Format_command::UINT) ?
-						va_arg(list, unsigned long) : va_arg(list, long);
+						(long long)va_arg(list, unsigned long) : va_arg(list, long);
 					break;
 
 				case Format_command::SIZE_T:
@@ -287,7 +287,7 @@ void Console::vprintf(const char *format, va_list list)
 				case Format_command::DEFAULT:
 
 					numeric_arg = (cmd.type == Format_command::UINT) ?
-						va_arg(list, unsigned int) : va_arg(list, int);
+						(long long)va_arg(list, unsigned int) : va_arg(list, int);
 					break;
 			}
 		}
